chore(NA): move missing canvas plugin tests out of __tests__ folder (#87898)
