+++ x-pack/plugins/canvas/shareable_runtime/components/footer/scrubber.test.tsx
@@ -6,14 +6,14 @@
 
 import { mount } from 'enzyme';
 import React from 'react';
-import { JestContext } from '../../../test/context_jest';
-import { Scrubber } from '../scrubber';
+import { JestContext } from '../../test/context_jest';
+import { Scrubber } from './scrubber';
 import {
   getScrubberSlideContainer as container,
   getRenderedElement as element,
-} from '../../../test/selectors';
+} from '../../test/selectors';
 
-jest.mock('../../../supported_renderers');
+jest.mock('../../supported_renderers');
 
 describe('<Scrubber />', () => {
   test('null workpad renders nothing', () => {
